pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Alabuga <None>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
django-elasticsearch-dsl = "^7.2.2"
elasticsearch-dsl = "^7.4.0"
djangorestframework = "^3.14.0"
django-elasticsearch-dsl-drf = "^0.22.5"
jwt = "^1.3.1"
django-cors-headers = "^3.13.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
